Fixes #9. This commit modifies the regex to match a given tag.
Previously, the regex matched only the first tag listed, so the
annotation was incorrect. Now, the user can specify the tag as an
argument (or use basic as default), and the tag will be included.

This isn't totally satisfactory, because I'd like to have all tags, and
add them as rows to the processed gtf file, but parsing the tags is too
hard and time consuming for right now. I'll add that as a new issue.